Make firmware a docker submodule to better indicate that the implementation is container-specific
